JIRATICKET-feat-Fixed the websocket issue
We were facing an issue where we werent able to send
the user details through the websocket. After some debugging
I found out that I was sending the details event before the websocket
connection was established. This meant all of those details were
skipped. We solved this issue by sending an acknowledgement to
the server once the websocket is connected. Only then would we
send the user details.

Also tried to synchronise data between websocket sent data and
session stored data in views.