JIRATICKET-feat-Implemented a file hashing system which prevents from uploading multiple same images.
Also implemented the complete profile viewing feature.
The hashing is a SHA-256 hashing which basically stores file hashes.
once the user uploads, first check if the same file has already been stored.
if not, we upload the file and send the user file hash as well as the public_id.